detached map layers and settings to separate files
